+++ src/main/java/com/mjuAppSW/joA/geography/block/repository/BlockJpaRepository.java
@@ -1,12 +1,13 @@
-package com.mjuAppSW.joA.geography.block;
+package com.mjuAppSW.joA.geography.block.repository;
 
+import com.mjuAppSW.joA.geography.block.Block;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface BlockRepository extends JpaRepository<Block, Long> {
+public interface BlockJpaRepository extends JpaRepository<Block, Long> {
 
     @Query("SELECT b FROM Block b WHERE b.blocker.id = :blockerId AND b.blocked.id = :blockedId")
     Optional<Block> findEqualBlock(@Param("blockerId") Long blockerId, @Param("blockedId") Long blockedId);
